move OlMap to /components
