+++ src/modules/olMap/components/OlMap.js
@@ -1,21 +1,21 @@
 import { html } from 'lit-html';
-import { MvuElement } from '../MvuElement';
+import { MvuElement } from '../../MvuElement';
 import olCss from 'ol/ol.css';
 import css from './olMap.css';
 import { Map as MapOl, View } from 'ol';
 import { defaults as defaultControls, ScaleLine } from 'ol/control';
 import { defaults as defaultInteractions, PinchRotate } from 'ol/interaction';
-import { removeLayer } from '../../store/layers/layers.action';
-import { changeLiveRotation, changeZoomCenterAndRotation } from '../../store/position/position.action';
-import { $injector } from '../../injection';
-import { updateOlLayer, toOlLayerFromHandler, registerLongPressListener, getLayerById } from './olMapUtils';
-import { setBeingDragged, setClick, setContextClick, setPointerMove } from '../../store/pointer/pointer.action';
-import { setBeingMoved, setMoveEnd, setMoveStart } from '../../store/map/map.action';
+import { removeLayer } from '../../../store/layers/layers.action';
+import { changeLiveRotation, changeZoomCenterAndRotation } from '../../../store/position/position.action';
+import { $injector } from '../../../injection';
+import { updateOlLayer, toOlLayerFromHandler, registerLongPressListener, getLayerById } from '../olMapUtils';
+import { setBeingDragged, setClick, setContextClick, setPointerMove } from '../../../store/pointer/pointer.action';
+import { setBeingMoved, setMoveEnd, setMoveStart } from '../../../store/map/map.action';
 import VectorSource from 'ol/source/Vector';
 import { Group as LayerGroup } from 'ol/layer';
-import { GeoResourceTypes } from '../../services/domain/geoResources';
-import { setFetching } from '../../store/network/network.action';
-import { emitNotification, LevelTypes } from '../../store/notifications/notifications.action';
+import { GeoResourceTypes } from '../../../services/domain/geoResources';
+import { setFetching } from '../../../store/network/network.action';
+import { emitNotification, LevelTypes } from '../../../store/notifications/notifications.action';
 
 const Update_Position = 'update_position';
 const Update_Layers = 'update_layers';
