+++ test/modules/olMap/components/OlMap.test.js
@@ -1,30 +1,30 @@
 /* eslint-disable no-undef */
-import { OlMap } from '../../../src/modules/olMap/OlMap';
+import { OlMap } from '../../../../src/modules/olMap/components/OlMap';
 import { fromLonLat } from 'ol/proj';
-import { TestUtils } from '../../test-utils.js';
-import { positionReducer } from '../../../src/store/position/position.reducer';
+import { TestUtils } from '../../../test-utils.js';
+import { positionReducer } from '../../../../src/store/position/position.reducer';
 import MapBrowserEventType from 'ol/MapBrowserEventType';
 import MapEventType from 'ol/MapEventType';
-import { $injector } from '../../../src/injection';
-import { layersReducer } from '../../../src/store/layers/layers.reducer';
-import { GeoResourceFuture, VectorGeoResource, VectorSourceType, WmsGeoResource } from '../../../src/services/domain/geoResources';
-import { addLayer, modifyLayer, removeLayer } from '../../../src/store/layers/layers.action';
-import { changeRotation, changeZoomAndCenter, fit } from '../../../src/store/position/position.action';
-import { simulateMapEvent, simulateMapBrowserEvent } from './mapTestUtils';
+import { $injector } from '../../../../src/injection';
+import { layersReducer } from '../../../../src/store/layers/layers.reducer';
+import { GeoResourceFuture, VectorGeoResource, VectorSourceType, WmsGeoResource } from '../../../../src/services/domain/geoResources';
+import { addLayer, modifyLayer, removeLayer } from '../../../../src/store/layers/layers.action';
+import { changeRotation, changeZoomAndCenter, fit } from '../../../../src/store/position/position.action';
+import { simulateMapEvent, simulateMapBrowserEvent } from '../mapTestUtils';
 import VectorLayer from 'ol/layer/Vector';
-import { pointerReducer } from '../../../src/store/pointer/pointer.reducer';
-import { mapReducer } from '../../../src/store/map/map.reducer';
+import { pointerReducer } from '../../../../src/store/pointer/pointer.reducer';
+import { mapReducer } from '../../../../src/store/map/map.reducer';
 import VectorSource from 'ol/source/Vector';
 import Event from 'ol/events/Event';
 import { Group as LayerGroup, Layer } from 'ol/layer';
-import { measurementReducer } from '../../../src/store/measurement/measurement.reducer';
-import { getDefaultLayerOptions } from '../../../src/modules/olMap/handler/OlLayerHandler';
-import { TEST_ID_ATTRIBUTE_NAME } from '../../../src/utils/markup';
-import { networkReducer } from '../../../src/store/network/network.reducer';
-import { createNoInitialStateMediaReducer } from '../../../src/store/media/media.reducer';
-import { setIsPortrait } from '../../../src/store/media/media.action';
-import { notificationReducer } from '../../../src/store/notifications/notifications.reducer';
-import { LevelTypes } from '../../../src/store/notifications/notifications.action';
+import { measurementReducer } from '../../../../src/store/measurement/measurement.reducer';
+import { getDefaultLayerOptions } from '../../../../src/modules/olMap/handler/OlLayerHandler';
+import { TEST_ID_ATTRIBUTE_NAME } from '../../../../src/utils/markup';
+import { networkReducer } from '../../../../src/store/network/network.reducer';
+import { createNoInitialStateMediaReducer } from '../../../../src/store/media/media.reducer';
+import { setIsPortrait } from '../../../../src/store/media/media.action';
+import { notificationReducer } from '../../../../src/store/notifications/notifications.reducer';
+import { LevelTypes } from '../../../../src/store/notifications/notifications.action';
 
 window.customElements.define(OlMap.tag, OlMap);
